# agentauth/docs/openapi.yaml
openapi: 3.1.0
info:
  title: AgentAuth API
  version: 0.7.0
  description: |
    **AgentAuth** is a production-ready authentication and identity service for AI agents.

    ## Features
    - JWT-based authentication
    - Rate limiting (100 req/15min general, 10 req/15min auth)
    - Webhooks for events
    - Activity logging
    - Permission-based access control
    - **Persona System** — define and manage an agent's digital soul (v0.7.0)
    - **ZKP Anonymous Verification** — prove identity without revealing credentials (v0.7.0)
    - **Anti-Drift Vault** — detect behavioral drift and auto-revoke rogue agents (v0.7.0)

    ## Authentication
    Most endpoints require an API key in the `X-Api-Key` header:
    ```
    X-Api-Key: ag_sk_your_api_key_here
    ```

    ## Rate Limits
    - **General endpoints**: 100 requests per 15 minutes
    - **Authentication endpoints**: 10 requests per 15 minutes

    Rate limit info is included in response headers:
    - `X-RateLimit-Limit`: Max requests allowed
    - `X-RateLimit-Remaining`: Requests remaining
    - `X-RateLimit-Reset`: Time until limit resets (Unix timestamp)

    ## Base URL
    - **Production**: `https://api.agentauth.dev`
    - **Local**: `http://localhost:3000`

  contact:
    name: AgentAuth Support
    email: umytbaynazarow754@gmail.com
    url: https://agentauth.dev
  license:
    name: MIT
    url: https://github.com/umyt-dev/Agent-Identity/blob/main/LICENSE
servers:
  - url: https://api.agentauth.dev
...
paths:
...
  /v1/agents/verify:
    post:
      summary: Verify agent credentials
      description: |
        Verify agent credentials (agent_id + API key) and receive agent details.

        **Rate limit:** 10 requests per 15 minutes
      operationId: verifyAgent
      tags:
        - Agents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentVerificationRequest'
            example:
              agent_id: ag_1a2b3c4d5e6f
              api_key: ag_sk_1a2b3c4d5e6f7g8h9i0j
      responses:
        '200':
          description: Agent verified successfully
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentVerificationResponse'
              example:
                valid: true
                agent_id: ag_1a2b3c4d5e6f
                name: CustomerSupportAgent
                tier: free
                permissions:
                  - read:messages
                  - write:responses
                status: active
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /v1/agents/{agent_id}/persona:
    post:
      summary: Register persona
      description: |
        Register a new persona (digital soul) for an agent. The persona is signed
        with HMAC-SHA256 using the agent's API key, creating a tamper-proof identity.

        Fires a `persona.created` webhook event.
      operationId: registerPersona
      tags:
        - Persona
      parameters:
        - $ref: '#/components/parameters/AgentId'
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonaRegistrationRequest'
            example:
              persona:
                version: '1.0.0'
                personality:
                  traits:
                    helpfulness: 0.9
                    formality: 0.7
                guardrails:
                  toxicity_threshold: 0.3
                  hallucination_tolerance: strict
                constraints:
                  forbidden_topics:
                    - politics
                    - religion
                  max_response_length: 2000
      responses:
        '201':
          description: Persona registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonaResponse'
              example:
                agent_id: ag_1a2b3c4d5e6f
                persona_version: '1.0.0'
                persona_hash: a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2
                created_at: '2026-02-01T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Persona already exists for this agent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Conflict
                message: Persona already registered. Use PUT to update.
        '413':
          description: Persona exceeds 10KB size limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Payload Too Large
                message: Persona validation failed
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      summary: Get persona
      description: |
        Retrieve the current persona for an agent. Supports ETag-based caching
        via `If-None-Match` header.

        Add `?include_prompt=true` to include the auto-generated system prompt.
      operationId: getPersona
      tags:
        - Persona
      parameters:
        - $ref: '#/components/parameters/AgentId'
        - name: include_prompt
          in: query
          description: Include generated system prompt template
          required: false
          schema:
            type: boolean
            default: false
        - name: If-None-Match
          in: header
          description: ETag from a previous response for conditional caching
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Persona retrieved
          headers:
            ETag:
              description: Hash-based ETag for caching
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonaFullResponse'
              example:
                agent_id: ag_1a2b3c4d5e6f
                persona:
                  version: '1.0.0'
                  personality:
                    traits:
                      helpfulness: 0.9
                      formality: 0.7
                persona_version: '1.0.0'
                persona_hash: a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2
                prompt: 'You are an AI agent with the following persona...'
        '304':
          description: Not modified (ETag match)
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update persona
      description: |
        Update an existing persona. The version is auto-bumped (minor) unless a
        higher explicit version is provided. Fires a `persona.updated` webhook.
      operationId: updatePersona
      tags:
        - Persona
      parameters:
        - $ref: '#/components/parameters/AgentId'
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonaRegistrationRequest'
      responses:
        '200':
          description: Persona updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonaResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/agents/{agent_id}/persona/verify:
    post:
      summary: Verify persona integrity
      description: |
        Re-compute the HMAC-SHA256 hash of the stored persona and compare it with
        the stored hash. Detects any tampering of the persona data at rest.
      operationId: verifyPersonaIntegrity
      tags:
        - Persona
      parameters:
        - $ref: '#/components/parameters/AgentId'
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Integrity check result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonaVerifyResponse'
              example:
                valid: true
                reason: Hash matches stored persona
                agent_id: ag_1a2b3c4d5e6f
                persona_version: '1.0.0'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/agents/{agent_id}/persona/history:
    get:
      summary: Get persona version history
      description: |
        Retrieve the version history of persona changes for an agent.
        Supports pagination and CSV export via `?format=csv`.
      operationId: getPersonaHistory
      tags:
        - Persona
      parameters:
        - $ref: '#/components/parameters/AgentId'
        - name: limit
          in: query
          description: Maximum entries to return
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of entries to skip
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: format
          in: query
          description: Response format
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        '200':
          description: Persona version history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonaHistoryResponse'
              example:
                history:
                  - id: 1
                    persona_version: '1.1.0'
                    persona_hash: a1b2c3d4...
                    changed_at: '2026-02-01T14:00:00.000Z'
                  - id: 0
                    persona_version: '1.0.0'
                    persona_hash: e5f6a1b2...
                    changed_at: '2026-02-01T12:00:00.000Z'
                total: 2
            text/csv:
              schema:
                type: string
              example: |
                version,hash,changed_at
                1.1.0,a1b2c3d4...,2026-02-01T14:00:00.000Z
                1.0.0,e5f6a1b2...,2026-02-01T12:00:00.000Z
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/agents/{agent_id}/persona/export:
    get:
      summary: Export persona bundle
      description: |
        Download a ZIP archive containing the persona JSON, metadata, and
        integrity hash. Useful for backup or importing to another agent.
      operationId: exportPersona
      tags:
        - Persona
      parameters:
        - $ref: '#/components/parameters/AgentId'
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Persona bundle ZIP
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/agents/{agent_id}/persona/import:
    post:
      summary: Import persona bundle
      description: |
        Import a persona from a previously exported bundle. The persona
        will be re-signed with the target agent's API key.
      operationId: importPersona
      tags:
        - Persona
      parameters:
        - $ref: '#/components/parameters/AgentId'
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                bundle:
                  type: string
                  format: binary
                  description: ZIP file from persona export
      responses:
        '201':
          description: Persona imported successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonaResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ─── ZKP Anonymous Verification ──────────────────────────────────────────────
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Agent API key
  parameters:
    AgentId:
      name: agent_id
      in: path
      description: Unique agent identifier
      required: true
      schema:
        type: string
        pattern: '^ag_[a-zA-Z0-9]+$'
      example: ag_1a2b3c4d5e6f
...
  headers:
    X-RateLimit-Limit:
      description: Maximum number of requests allowed in the current window
      schema:
        type: integer
      example: 10
    X-RateLimit-Remaining:
      description: Number of requests remaining in the current window
      schema:
        type: integer
      example: 7
    X-RateLimit-Reset:
      description: Unix timestamp when the rate limit resets
      schema:
        type: integer
      example: 1738329600
  schemas:
...
    AgentVerificationRequest:
      type: object
      required: [agent_id, api_key]
      properties:
        agent_id:
          type: string
        api_key:
          type: string
    AgentVerificationResponse:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
        agent_id:
          type: string
        name:
          type: string
        tier:
          type: string
          enum: [free, pro, enterprise]
        permissions:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, suspended]
        error:
          type: string
...
    PersonaObject:
      type: object
      required: [version]
      description: |
        The persona definition — an agent's "digital soul". Max size 10KB.
        See [Persona Schema Reference](/docs/persona-schema) for full details.
      properties:
        version:
          type: string
          description: Semantic version (e.g. "1.0.0")
          pattern: '^\d+\.\d+\.\d+.*$'
        personality:
          type: object
          properties:
            traits:
              type: object
              description: Key-value pairs. Numbers must be 0-1; strings must be from the allowed traits list.
              additionalProperties:
                oneOf:
                  - type: number
                    minimum: 0
                    maximum: 1
                  - type: string
            assistantAxis:
              type: array
              items:
                type: object
            neuralVectors:
              type: object
        guardrails:
          type: object
          properties:
            toxicity_threshold:
              type: number
              minimum: 0
              maximum: 1
            hallucination_tolerance:
              type: string
              enum: [strict, moderate, lenient]
            source_citation_required:
              type: boolean
        constraints:
          type: object
          properties:
            forbidden_topics:
              type: array
              items:
                type: string
            required_disclaimers:
              type: array
              items:
                type: string
            allowed_actions:
              type: array
              items:
                type: string
            blocked_actions:
              type: array
              items:
                type: string
            max_response_length:
              type: integer
              minimum: 1
      additionalProperties: true
    PersonaRegistrationRequest:
      type: object
      required: [persona]
      properties:
        persona:
          $ref: '#/components/schemas/PersonaObject'
    PersonaResponse:
      type: object
      properties:
        agent_id:
          type: string
        persona_version:
          type: string
        persona_hash:
          type: string
          description: HMAC-SHA256 hex digest (64 chars)
        created_at:
          type: string
          format: date-time
    PersonaFullResponse:
      type: object
      properties:
        agent_id:
          type: string
        persona:
          $ref: '#/components/schemas/PersonaObject'
        persona_version:
          type: string
        persona_hash:
          type: string
        prompt:
          type: string
          description: Auto-generated system prompt (only with include_prompt=true)
    PersonaVerifyResponse:
      type: object
      required: [valid, reason]
      properties:
        valid:
          type: boolean
        reason:
          type: string
        agent_id:
          type: string
        persona_version:
          type: string
    PersonaHistoryEntry:
      type: object
      properties:
        id:
          type: integer
        persona_version:
          type: string
        persona_hash:
          type: string
        changed_at:
          type: string
          format: date-time
    PersonaHistoryResponse:
      type: object
      properties:
        history:
          type: array
          items:
            $ref: '#/components/schemas/PersonaHistoryEntry'
        total:
          type: integer

    # ─── ZKP Schemas ─────────────────────────────────────────────────────────
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
  responses:
    BadRequest:
      description: Bad request - validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Validation failed
            message: Invalid input data
            details:
              - field: email
                message: Must be a valid email address
    Unauthorized:
      description: Unauthorized - invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Invalid agent credentials
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not Found
            message: Resource not found
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          schema:
            type: integer
            example: 0
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Rate Limit Exceeded
            message: Too many requests. Please try again later.
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal Server Error
            message: An unexpected error occurred